requirements.txt
openai-whisper
openai
python-dotenv
pyaudio